split animation cycle classes into multiple files
