+++ animation-cycle.cpp
@@ -1,62 +1,5 @@
 #include "animation-cycle.hpp"
-#include "colors.hpp"
 
-Animation::Animation()
-    : next_(nullptr)
-{}
-
-Animation::~Animation() {
-    if (next_ != nullptr) {
-        delete next_;
-    }
-}
-
-Animation* Animation::next(){
-    return next_;
-}
-
-void Animation::set_next(Animation *next) {
-    next->next_ = next_;
-    next_ = next;
-}
-
-
-StaticRing::StaticRing(const Ring_t& ring)
-    : Animation(),
-    ring_(&ring)
-{}
-
-bool StaticRing::is_finished(unsigned int ms_since_animation_start) {
-    return true;
-}
-
-uint32_t StaticRing::get_color(uint8_t pixel_index, unsigned int ms_since_animation_start, const Ring_t& base_color) {
-    return (*ring_)[pixel_index];
-}
-
-Wait::Wait(unsigned int ms_wait)
-    : Animation(),
-    ms_wait_(ms_wait)
-{}
-
-bool Wait::is_finished(unsigned int ms_since_animation_start) {
-    return ms_since_animation_start > ms_wait_;
-}
-
-uint32_t Wait::get_color(uint8_t pixel_index, unsigned int ms_since_animation_start, const Ring_t& base_color) {
-    return base_color[pixel_index];
-}
-
-AnimationCycleConfig::AnimationCycleConfig(Animation* first_animation)
-: last_animation_(first_animation)
-{}
-
-AnimationCycleConfig& AnimationCycleConfig::then(Animation* animation)
-{
-    last_animation_->set_next(animation);
-    last_animation_ = animation;
-    return *this;
-}
 
 AnimationCycle::AnimationCycle(Adafruit_NeoPixel& neopixels, int frames_per_second)
     : neopixels_(neopixels),
@@ -67,10 +10,11 @@ AnimationCycle::~AnimationCycle() {
     delete first_animation_;
 }
 
-AnimationCycleConfig AnimationCycle::set_up(const Ring_t& start_color) {
-    first_animation_ = new StaticRing(start_color);
+AnimationCycleBuilder AnimationCycle::build_from(const Ring_t& start_color) {
+    AnimationCycleBuilder builder;
+    first_animation_ = builder.init(start_color);
     animation_ = first_animation_;
-    return AnimationCycleConfig(first_animation_);
+    return builder;
 }
 
 void AnimationCycle::start() {
